+++ common/src/debug/res/drawable/custom_surface2_radius12.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_focused="false" android:state_pressed="false">
+        <shape>
+            <solid android:color="@color/surface2"/>
+            <corners android:radius="12dp" />
+        </shape>
+    </item>
+</selector>
