+++ feature/home/src/main/java/com/dida/home/HomeBindingAdapters.kt
@@ -1,10 +1,9 @@
-package com.dida.android.presentation.views.nav.home
+package com.dida.home
 
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.dida.android.presentation.adapter.home.*
-import com.dida.android.util.UiState
-import com.dida.android.util.successOrNull
+import com.dida.common.util.UiState
+import com.dida.common.util.successOrNull
 import com.dida.domain.model.nav.home.*
 
 @BindingAdapter("hotsItem")
