+++ feature/home/src/main/java/com/dida/home/HomeViewModel.kt
@@ -1,8 +1,9 @@
-package com.dida.android.presentation.views.nav.home
+package com.dida.home
 
 import com.dida.android.util.NftActionHandler
 import com.dida.android.util.UiState
 import com.dida.common.base.BaseViewModel
+import com.dida.common.util.UiState
 import com.dida.domain.flatMap
 import com.dida.domain.model.nav.home.Home
 import com.dida.domain.model.nav.home.SoldOut
