+++ feature/home/src/main/java/com/dida/home/adapter/SoldOutAdapter.kt
@@ -1,4 +1,4 @@
-package com.dida.android.presentation.adapter.home
+package com.dida.home.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -7,20 +7,20 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.dida.android.R
-import com.dida.android.databinding.HolderCollectionBinding
-import com.dida.android.presentation.views.nav.home.HomeActionHandler
-import com.dida.domain.model.nav.home.Collection
+import com.dida.android.databinding.HolderSoldoutBinding
+import com.dida.home.HomeActionHandler
+import com.dida.domain.model.nav.home.SoldOut
 
-class CollectionAdapter(
+class SoldOutAdapter(
     private val eventListener: HomeActionHandler
-) : ListAdapter<Collection, CollectionAdapter.ViewHolder>(CollectionItemDiffCallback){
+): ListAdapter<SoldOut, SoldOutAdapter.ViewHolder>(SoldOutItemDiffCallback) {
 
     init { setHasStableIds(true) }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val viewDataBinding: HolderCollectionBinding = DataBindingUtil.inflate(
+        val viewDataBinding: HolderSoldoutBinding = DataBindingUtil.inflate(
             LayoutInflater.from(parent.context),
-            R.layout.holder_collection,
+            R.layout.holder_soldout,
             parent,
             false
         )
@@ -32,20 +32,19 @@ class CollectionAdapter(
         holder.bind(getItem(position))
     }
 
-    class ViewHolder(private val binding: HolderCollectionBinding) :
+    class ViewHolder(private val binding: HolderSoldoutBinding) :
         RecyclerView.ViewHolder(binding.root) {
-
-        fun bind(item: Collection) {
+        fun bind(item: SoldOut) {
             binding.holderModel = item
             binding.executePendingBindings()
         }
     }
 
-    internal object CollectionItemDiffCallback : DiffUtil.ItemCallback<Collection>() {
-        override fun areItemsTheSame(oldItem: Collection, newItem: Collection) =
+    internal object SoldOutItemDiffCallback : DiffUtil.ItemCallback<SoldOut>() {
+        override fun areItemsTheSame(oldItem: SoldOut, newItem: SoldOut) =
             oldItem == newItem
 
-        override fun areContentsTheSame(oldItem: Collection, newItem: Collection) =
+        override fun areContentsTheSame(oldItem: SoldOut, newItem: SoldOut) =
             oldItem.equals(newItem)
     }
 }
