+++ feature/home/src/main/res/layout/fragment_home.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="vm"
-            type="com.dida.android.presentation.views.nav.home.HomeViewModel" />
+            type="com.dida.home.HomeViewModel" />
     </data>
